% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacheR.R
\name{saveAnswer}
\alias{saveAnswer}
\title{Save the question's answer}
\usage{
saveAnswer(x, dir_path = "answer")
}
\arguments{
\item{x}{A character}

\item{dir_path}{A character}
}
\description{
Save the question's answer
}
\examples{
fileName <- saveAnswer('this is a answer', tempdir())
loadAnswer(fileName, tempdir())
}

